Shields now block it based on their block chance

+++ code/game/objects/items/weapons/humanball.dm
@@ -14,15 +14,17 @@
 /mob/living/carbon/human/hitby(atom/movable/AM, blocked = 0)
 
 	..()
-	if(istype(AM,/obj/item/humanpokeball))
-		unEquip(wear_mask) //Will probably screw over people who need to use internals.
-		unEquip(wear_suit) // Might screw over people wearing suits in space.
-		unEquip(glasses)
-		equipOutfit(/datum/outfit/pokeball)
-		apply_effect(2, WEAKEN) //Straight jackets do not slow movement, this should give the user enough time to run over and snag the guy.
 
-		visible_message("<span class='danger'>Success! The [src] was caught!</span>", \
-					"<span class='userdanger'>The device wraps restraints around your body and pulls them tight!</span>")
-		qdel(AM)
+	if(blocked <= 0)
+		if(istype(AM,/obj/item/humanpokeball))
+			unEquip(wear_mask) //Will probably screw over people who need to use internals.
+			unEquip(wear_suit) // Might screw over people wearing suits in space.
+			unEquip(glasses)
+			equipOutfit(/datum/outfit/pokeball)
+			apply_effect(2, WEAKEN) //Straight jackets do not slow movement, this should give the user enough time to run over and snag the guy.
+
+			visible_message("<span class='danger'>Success! The [src] was caught!</span>", \
+						"<span class='userdanger'>The device wraps restraints around your body and pulls them tight!</span>")
+			qdel(AM)
